% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_CBS}
\alias{population_CBS}
\title{Sudan population projection 2018 by locality from Sudan Central Bureau of
Statistics}
\format{
A data.frame of 3 columns and 189 rows:
\describe{
  \item{\code{state}}{State name}
  \item{\code{locality}}{Locality name}
  \item{\code{pop}}{Population}
}
}
\source{
Sudan Central Bureau of Statistics
}
\usage{
population_CBS
}
\description{
Sudan population projection 2018 by locality from Sudan Central Bureau of
Statistics
}
\examples{
population_CBS

}
\keyword{datasets}
