% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_S3M}
\alias{population_S3M}
\title{Sudan population projection 2018 by locality from Sudan Central Bureau of
Statistics and UNICEF and Ministry of Health including 3 IDP camps for use
in S3M analysis}
\format{
A data.frame of 3 columns and 192 rows:
\describe{
  \item{\code{state}}{State name}
  \item{\code{locality}}{Locality name}
  \item{\code{pop}}{Population}
}
}
\source{
Sudan Central Bureau of Statistics, UNICEF and Ministry of Health
}
\usage{
population_S3M
}
\description{
Sudan population projection 2018 by locality from Sudan Central Bureau of
Statistics and UNICEF and Ministry of Health including 3 IDP camps for use
in S3M analysis
}
\examples{
population_S3M

}
\keyword{datasets}
