% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-sd_get_pop.R
\name{sd_get_page_locality}
\alias{sd_get_page_locality}
\title{Exract per page locality population tables from a PDF copy of the UNFPA's
Sudan population projections from 2009-2020}
\usage{
sd_get_page_locality(
  page,
  startRow,
  period,
  labs = c("total", "male", "female"),
  sep = "_"
)
}
\arguments{
\item{page}{An object representing one page text from the UNFPA's Sudan
population projections for a locality. This is extracted using the
\code{pdf_text()} function from the \code{pdftools} package}

\item{startRow}{Numeric value indicating the starting point of actual data
from the tables}

\item{period}{A range of years in 4-year groups starting from 2009 up to
2020}

\item{labs}{Labels for the population data columns for each reported year;
Default to c("total", "male", "female")}

\item{sep}{Separator to use when labeling columns using \code{labs} and
reporting year}
}
\value{
A data.frame of population data from the given page and specified
years
}
\description{
Exract per page locality population tables from a PDF copy of the UNFPA's
Sudan population projections from 2009-2020
}
