% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-sd_get_pop.R
\name{sd_get_pop_locality}
\alias{sd_get_pop_locality}
\title{Exract locality population data from a set of pages for a specific state
from UNFPA's Sudan population projections for 2009-2020}
\usage{
sd_get_pop_locality(corpus, state, localities)
}
\arguments{
\item{corpus}{An object representing text data extracted from the UNPFA
Sudan population projections for a specific state. This is extracted using
the \code{pdf_text()} function from the \code{pdftools} package}

\item{state}{Name of state to which current data is for}

\item{localities}{A vector of localities name for the given state}
}
\value{
A data.frame of population data for localities in a given state
}
\description{
Exract locality population data from a set of pages for a specific state
from UNFPA's Sudan population projections for 2009-2020
}
