% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-tidy_pop.R
\name{sd_tidy_pop}
\alias{sd_tidy_pop}
\title{Tidy up population projections data from UNFPA}
\usage{
sd_tidy_pop()
}
\value{
A tibble containing 6 columns and 110808 rows that is a longer
  format version of the UNFPA population dataset
}
\description{
A convenience wrapper to tidyr::pivot_longer to switch population projections
from UNFPA to long format.
}
\examples{
sd_tidy_pop()

}
