% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{Map of states of Sudan}
\format{
A SpatialPolygonsDataFrame with 18 features and 10 fields

| **Variables** | **Description** |
| *OBJECTID_1* | Object identifier |
| *POP_2008* | Population in 2008 |
| *POP_2018* | Population in 2018 |
| *POP2019* | Population in 2019 |
| *State_En* | Name of state (English) |
| *Shape_Le_2* | Shape size |
| *Area_1* | Area size |
| *Code_1* | Area code |
| *stateID* | State identifier |
}
\source{
Sudan Central Bureau of Statistics
}
\usage{
states
}
\description{
Map of states of Sudan
}
\keyword{datasets}
