% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{localities}
\alias{localities}
\title{Map of localities of Sudan}
\format{
A SpatialPolygonsDataFrame with 188 features and 11 fields.

| **Variables** | **Description** |
| *OBJECTID_1* | Object identifier 1 |
| *OBJECTID* | Object identifier |
| *Toto_Pop* | Total population |
| *Name_Engli* | Name of state (English) |
| *Locality_G* | Locality code |
| *Shape_Le_2* | Shape size |
| *Shape_Area* | Shape area |
| *stateID* | State identifier |
| *localityID* | Locality identifier |
}
\source{
Sudan Central Bureau of Statistics
}
\usage{
localities
}
\description{
Map of localities of Sudan
}
\keyword{datasets}
