% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{Map of states of Sudan}
\format{
A SpatialPolygonsDataFrame with 18 features and 10 fields\tabular{ll}{
   \strong{Variables} \tab \strong{Description} \cr
   \emph{OBJECTID_1} \tab Object identifier \cr
   \emph{POP_2008} \tab Population in 2008 \cr
   \emph{POP_2018} \tab Population in 2018 \cr
   \emph{POP2019} \tab Population in 2019 \cr
   \emph{State_En} \tab Name of state (English) \cr
   \emph{Shape_Le_2} \tab Shape size \cr
   \emph{Area_1} \tab Area size \cr
   \emph{Code_1} \tab Area code \cr
   \emph{stateID} \tab State identifier \cr
}
}
\source{
Sudan Central Bureau of Statistics
}
\usage{
states
}
\description{
Map of states of Sudan
}
\keyword{datasets}
